requirements.txt
langchain
openai
python-dotenv
pinecone-client
unstructured
pdf2image
pdfminer.six
tiktoken